+++ climada/util/files_handler.py
@@ -25,17 +25,301 @@ __all__ = [
 ]
 
 import glob
+import hashlib
 import logging
 import math
+import time
 import urllib
+from datetime import datetime, timezone
 from pathlib import Path
 
 import requests
+from peewee import (
+    AutoField,
+    CharField,
+    DateTimeField,
+    IntegerField,
+    IntegrityError,
+    Model,
+    SqliteDatabase,
+)
+from playhouse.migrate import SqliteMigrator, migrate
 from tqdm import tqdm
 
 from climada.util.config import CONFIG
 
 LOGGER = logging.getLogger(__name__)
+
+DB = SqliteDatabase(Path(CONFIG.data_api.cache_db.str()).expanduser())
+
+
+class Download(Model):
+    """Database entry keeping track of downloaded files from the CLIMADA data API"""
+
+    # If you do not specify a primary key, Peewee will automatically create an auto-incrementing
+    # primary key named “id”.
+    url = CharField()
+    path = CharField(unique=True)
+    startdownload = DateTimeField()
+    enddownload = DateTimeField(null=True)
+    timestamp = IntegerField(null=True)
+    filesize = IntegerField(null=True)
+    checksum = CharField(null=True)
+
+    class Meta:
+        """SQL database and table definition."""
+
+        database = DB
+
+    class Failed(Exception):
+        """The download failed for some reason."""
+
+
+def init_DB():
+    DB.connect()
+    DB.create_tables([Download])
+    try:
+        migrator = SqliteMigrator(DB)
+        migrate(
+            migrator.add_column("download", "checksum", CharField(null=True)),
+            migrator.add_column("download", "timestamp", IntegerField(null=True)),
+            migrator.add_column("download", "filesize", IntegerField(null=True)),
+        )
+    except Exception as exc:
+        print(exc)
+    DB.close()
+
+
+init_DB()
+
+HASH_FUNCS = {
+    "md5": hashlib.md5,
+    "sha1": hashlib.sha1,
+}
+
+
+def file_checksum(filename, hash_func):
+    hf = HASH_FUNCS[hash_func]()
+    ba = bytearray(128 * 1024)
+    mv = memoryview(ba)
+    with open(filename, "rb", buffering=0) as f:
+        for n in iter(lambda: f.readinto(mv), 0):
+            hf.update(mv[:n])
+    hashsum = hf.hexdigest()
+    return f"{hash_func}:{hashsum}"
+
+
+class Downloader:
+
+    MAX_WAITING_PERIOD = 6
+    DOWNLOAD_TIMEOUT = 3600
+
+    def __init__(self, checksum=None):
+        """Downloader
+
+        Parameters
+        ----------
+        checksum : {'md5', 'sha1}, optional
+            if not None, the checksum of downloaded files are being recorded.
+            Default: None
+        """
+        self.chunk_size = CONFIG.data_api.chunk_size.int()
+        self.checksum = checksum
+
+    def _tracked_download(self, remote_url, local_path):
+        if local_path.is_dir():
+            raise ValueError(
+                "tracked download requires a path to a file not a directory"
+            )
+        path_as_str = str(local_path.absolute())
+        try:
+            dlf = Download.create(
+                url=remote_url, path=path_as_str, startdownload=datetime.utcnow()
+            )
+        except IntegrityError as ierr:
+            dlf = Download.get(
+                Download.path == path_as_str
+            )  # path is the table's one unique column
+            if not Path(path_as_str).is_file():  # in case the file has been removed
+                dlf.delete_instance()  # delete entry from database
+                return self._tracked_download(remote_url, local_path)  # and try again
+            if dlf.url != remote_url:
+                raise RuntimeError(
+                    f"this file ({path_as_str}) has been downloaded from another url ({dlf.url}),"
+                    " possibly because it belongs to a dataset with a recent version update."
+                    " Please remove the file or purge the entry from data base before trying again"
+                ) from ierr
+            return dlf
+        try:
+            # the actual download from url, using requests.get method
+            with requests.get(
+                remote_url, stream=True, timeout=Downloader.DOWNLOAD_TIMEOUT
+            ) as stream:
+                stream.raise_for_status()
+                with open(local_path, "wb") as dump:
+                    for chunk in stream.iter_content(chunk_size=self.chunk_size):
+                        dump.write(chunk)
+            # update the db entry
+            dlf.enddownload = datetime.now(timezone.utc)
+            dlf.timestamp = int(local_path.stat().st_mtime)
+            dlf.filesize = local_path.stat().st_size
+            if self.checksum:
+                dlf.checksum = file_checksum(local_path, self.checksum)
+            dlf.save()
+        except Exception:
+            dlf.delete_instance()
+            raise
+        return Download.get(Download.path == path_as_str)
+
+    def download(self, url, target_dir, file_name, integrity_check=None):
+        """Download a file if it is not already present at the target destination.
+
+        Parameters
+        ----------
+        local_path : Path
+            target destination,
+            if it is a directory the original filename (fileinfo.filen_name) is kept
+        fileinfo : FileInfo
+            file object as retrieved from the data api
+        integrity_check : function(path) -> (), optional
+            the method that is used to check the integrity of the already downloaded file.
+            expected to raise a `Download.Failed` exception if it wasn't successfull.
+            Default: `Downloader.file_unchanged_since`.
+
+        Returns
+        -------
+        Path
+            the path to the downloaded file
+
+        Raises
+        ------
+        Exception
+            when number of retries was exceeded or when a download is already running
+        """
+        return self._download_file(
+            url=url,
+            target_dir=target_dir,
+            file_name=file_name,
+            check_method=integrity_check or self.file_unchanged_since,
+            retries=3,
+        )
+
+    def _download_file(self, url, target_dir, file_name, check_method, retries):
+        """Download a file if it is not already present at the target destination.
+
+        Parameters
+        ----------
+        local_path : Path
+            target destination,
+            if it is a directory the original filename (fileinfo.filen_name) is kept
+        fileinfo : FileInfo
+            file object as retrieved from the data api
+        check : function, optional
+            how to check download success, by default checksize
+        retries : int, optional
+            how many times one should retry in case of failure, by default 3
+
+        Returns
+        -------
+        Path
+            the path to the downloaded file
+
+        Raises
+        ------
+        Exception
+            when number of retries was exceeded or when a download is already running
+        """
+        try:
+            local_path = Path(target_dir, file_name)
+            downloaded = self._tracked_download(remote_url=url, local_path=local_path)
+            if not downloaded.enddownload:
+                raise Download.Failed(
+                    f"A download of {url} via the climada.util.files_handler.Downloader has been"
+                    " requested before. Either it is still in progress or the process got"
+                    " interrupted. In the former case just wait until the download has finished"
+                    " and try again, in the latter run `Downloader.purge_cache_db(Path('"
+                    f"{local_path}'))` from Python. If unsure, check your internet connection,"
+                    " wait for as long as it takes to download a file of the given size and try"
+                    " again. If the problem persists, purge the cache db with said call."
+                )
+            try:
+                check_method(local_path)
+            except Download.Failed as dlf:
+                local_path.unlink(missing_ok=True)
+                self.purge_cache_db(local_path)
+                raise dlf
+            return local_path
+        except Download.Failed as dle:
+            if retries < 1:
+                raise dle
+            LOGGER.warning("Download failed: %s, retrying...", dle)
+            time.sleep(Downloader.MAX_WAITING_PERIOD / retries)
+            return self._download_file(
+                url=url,
+                target_dir=target_dir,
+                file_name=file_name,
+                check_method=check_method,
+                retries=retries - 1,
+            )
+
+    @staticmethod
+    def purge_cache_db(local_path):
+        """Removes entry from the sqlite database that keeps track of files downloaded by
+        `_tracked_download`. This may be necessary in case a previous attempt has failed
+        in an uncontroled way (power outage or the like).
+
+        Parameters
+        ----------
+        local_path : Path
+            target destination
+        fileinfo : FileInfo
+            file object as retrieved from the data api
+        """
+        dlf = Download.get(Download.path == str(local_path.absolute()))
+        dlf.delete_instance()
+
+    def file_unchanged_since(self, local_path):
+        """default method for checking file integrity
+        basically checks whether the file has changed since the download
+
+        Parameters
+        ----------
+        local_path : Path
+            the file whose integrity is being scrutinized
+
+        Raises
+        ------
+        Download.Failed
+            if the file is not there or doesn't have the same size anymore
+            or the same timestamp, or the same checksum in case checksums are recorded
+        """
+        if not local_path.is_file():
+            raise Download.Failed(f"{str(local_path)} is not a file")
+        dlf = Download.get(Download.path == str(local_path.absolute()))
+        if local_path.stat().st_size != dlf.filesize:
+            raise Download.Failed(
+                f"{str(local_path)} has the wrong size:"
+                f"{local_path.stat().st_size} instead of {dlf.filesize}"
+            )
+        if self.checksum:
+            csm, hsh = dlf.checksum.split(":")
+            found = file_checksum(local_path, csm)
+            if hsh != found:
+                raise Download.Failed(
+                    f"{str(local_path)} has changed, checksums differ: "
+                    f"{csm}:{hsh}, {csm}:{found}"
+                )
+        else:  # if the checksum is still the same, there is no point comparing timestamps!
+            # even if it has changed, the file has _certainly_ just been touched.
+            if int(local_path.stat().st_mtime) != dlf.timestamp:
+                expected = datetime.fromtimestamp(dlf.timestamp).isoformat()
+                found = datetime.fromtimestamp(
+                    int(local_path.stat().st_mtime)
+                ).isoformat()
+                raise Download.Failed(
+                    f"{str(local_path)} has been modified: timestamp changed from"
+                    f"{expected} to {found}"
+                )
 
 
 class DownloadProgressBar(tqdm):
@@ -202,20 +486,3 @@ def get_file_names(file_name):
                 raise ValueError(f'cannot find the file "{pattern}"')
             file_list.extend(extension)
     return file_list
-
-
-def get_extension(file_name):
-    """Get file without extension and its extension (e.g. ".nc", ".grd.gz").
-
-    Parameters
-    ----------
-    file_name : str
-        file name (with or without path)
-
-    Returns
-    -------
-    str, str
-    """
-    file_path = Path(file_name)
-    cuts = file_path.name.split(".")
-    return str(file_path.parent.joinpath(cuts[0])), "".join(file_path.suffixes)